+++ swap/src/protocol/bob/event_loop.rs
@@ -3,7 +3,7 @@ use crate::network::quote::BidQuote;
 use crate::network::{encrypted_signature, spot_price};
 use crate::protocol::bob::{Behaviour, OutEvent, State0, State2};
 use crate::{bitcoin, monero};
-use anyhow::{Context, Result};
+use anyhow::{bail, Context, Result};
 use futures::future::{BoxFuture, OptionFuture};
 use futures::{FutureExt, StreamExt};
 use libp2p::request_response::{RequestId, ResponseChannel};
@@ -261,11 +261,22 @@ impl EventLoopHandle {
     }
 
     pub async fn request_spot_price(&mut self, btc: bitcoin::Amount) -> Result<monero::Amount> {
-        Ok(self
+        let response = self
             .spot_price
             .send_receive(spot_price::Request { btc })
-            .await?
-            .xmr)
+            .await?;
+
+        match (response.xmr, response.error) {
+            (Some(xmr), None) => Ok(xmr),
+            (_, Some(error)) => {
+                bail!(error);
+            }
+            (None, None) => {
+                bail!(
+                    "Unexpected response for spot-price request, neither price nor error received"
+                );
+            }
+        }
     }
 
     pub async fn request_quote(&mut self) -> Result<BidQuote> {