+++ swap/tests/harness/mod.rs
@@ -234,6 +234,7 @@ fn start_alice(
         db,
         FixedRate::default(),
         bitcoin::Amount::ONE_BTC,
+        false,
     )
     .unwrap();